Rework Battle and Rotation Fixes
Reworks battles to be more like other frames and use basic text rotation to achieve layout instead of complicated logic. These no longer save sideways either and can now be used with margin. Also adds the ability to rotate set symbols from the set symbol tab and you can set the rotation in the packs. This also fixes how art rotation is handled so that the art rotates from the center of the art rather than a corner of the canvas and you can now set art rotation in the packs.