Fixed Outline on Text
This fixes how outline is drawn for text so that it is all drawn before the fill so we don't get overlap of outlines on previous lines of text. This also adjusts the Borderless Source Material Text so that it is closer in spacing now to accurate since we don't have to worry about text overlap with the outlines. Also fixes an issue we had with the replacement system for line breaks on the Borderless Source Material when {indent} was used and not ended beofre the next {lns} replacement